pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ador"
version = "0.1.0"
description = "Augmented DOcument Reader — financial document entity extraction (CMI A&I PoC)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "CMI Architecture & Innovation" }]
license = { text = "Proprietary" }
dependencies = [
    "pydantic>=2.6",
    "python-docx>=1.1",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "typer>=0.12",
]

[project.optional-dependencies]
ner = [
    "transformers>=4.40",
    "torch>=2.2",
    "spacy>=3.7",
]
llm = [
    "anthropic>=0.25",
    "pypdf>=4.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "httpx>=0.27",
]

[project.scripts]
ador = "ador.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "N", "RUF"]

[tool.ruff.lint.per-file-ignores]
# FastAPI and Typer require callables (Depends, File, typer.Argument/Option)
# as argument defaults — B008 is a false positive for these frameworks.
"src/ador/api/*.py" = ["B008"]
"src/ador/cli.py" = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
